pyproject.toml
[project]
name = "ai-pr-reviewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.2",
    "openai>=1.109.1",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
]
